Handle abstract element substitution groups.
When processing a substitution group based on an abstract element, the
generated code would not compile if the element was nested within a
sequence that was nested under a choice element. Loading whether an
element is abstract and only providing substitution group elements that
are not abstract solves this problem.